+++ rules/react/react.js
@@ -70,7 +70,7 @@ const legacyReactRules = {
 export default {
   plugins: {
     ...eslintReact.configs["recommended-typescript"].plugins,
-    "hoseop-react": legacyReactRules,
+    "broccoil-react": legacyReactRules,
     "react-hooks": reactHooks,
     "react-refresh": reactRefresh,
   },
@@ -103,7 +103,7 @@ export default {
     "@eslint-react/jsx-shorthand-fragment": "warn",
     "@eslint-react/no-useless-fragment": "warn",
     "@eslint-react/dom/no-script-url": "warn",
-    "hoseop-react/no-unsafe-target-blank": "warn",
+    "broccoil-react/no-unsafe-target-blank": "warn",
     "@eslint-react/no-unused-state": "warn",
     // Hooks 규칙 (error)
     "react-hooks/rules-of-hooks": "error",